Convert Connection to trait and make lib no_std by default
This is a mostly complete initial implementation of the conversion
of the Connection struct into a generic trait, as well as the
conversion of various other parts of the library to make use of
Connection implementers, rather than a hard-coded Connection
struct. Two optional implementations, TCPConnection and
TLSConnection are provided when features are enabled. By default,
a no_std user of the library must manually implement the
Connection trait on an object of their choosing and pass this to
Request.send().

In addition to the conversion of the Connection, the lib is now
no_std by default. When built with `cargo build`, the default
feature builds with no std-lib features, but also provides no
implementations for the Connection trait.