Fix test cases to match new behaviour and add new examples of trait
The new Connection trait introduces changes in usage for the library.
Each request requires the instantiation of a dedicated Connection
instance which is consumed during send(). Both the unit and
doc-tests have been updated to reflect this new functionality.

Additionally, two new examples have been provided. The first shows
the implementation of a Connection in a no_std environment. While
it uses std features for simplicity (such as println! and Vec),
it is mainly intended to demonstrate the required trait
implementations in a no_std environment, when the convenience
wrappers are not available for std types.

The std_conn example closely matches the nostd_conn example, but
demonstrates CoreReader usage to wrap std::io::Read instances
to automatically implement the required CoreRead trait for
the HttpStream enum.